+++ baseCompressAlgs/bbc.js
