+++ baseCompressAlgs/compressions.js
@@ -1,0 +1,3 @@
+export { wahCompress } from './wah.js'
+export { valCompress } from './val.js'
+export { bbcCompress } from './bbc.js'
