+++ baseCompressAlgs/data/animals
