+++ baseCompressAlgs/data/animals_BBC_8
