+++ baseCompressAlgs/data/animals_WAH_16
