+++ baseCompressAlgs/data/animals_WAH_32
