+++ baseCompressAlgs/data/animals_WAH_4
