+++ baseCompressAlgs/data/animals_WAH_64
