+++ baseCompressAlgs/data/animals_WAH_8
