+++ baseCompressAlgs/data/animals_sorted
