+++ baseCompressAlgs/data/animals_sorted_BBC_8
