+++ baseCompressAlgs/data/animals_sorted_WAH_64
