--- compressionAlgs/package.json
@@ -1,5 +1,0 @@
-{
-  "name": "my-project",
-  "version": "1.0.0",
-  "type": "module"
-}
